base commit
Set up a test suite to test out features. Able to add in parts of the url into the database

Finished basic functionality. Able to take in the URL and then spit it out when that link got posted again

Refactor the code so the async function is clearer. Double links in message fringe situation are dealt with

moved server name, password and driver url to the config
,simplified regex and actually close the session after it's done

Squashed commits due to security concern